+++ packages/superposition/src/index.ts
@@ -3,4 +3,3 @@ export * from './Error';
 export * from './Interface';
 export * from './Schrodinger';
 export * from './Superposition';
-export * from './SuperpositionInternal';